This fix allows the fileIO test to work with the PVM tests.
This fix allows the fileIO test to work with the PVM tests. First, it contains
a hack to clean up the environment from previous failed runs, but the hack is
UNIX specific: system call with touch /tmp/Chplpvmtmp && rm -rf /tmp/*pvm* &&
killall -q -9 pvmd3, which removes all tmp files associated with a PVM run and
kills the daemon silently (in the case it isn't running, we don't want the
message telling us the files aren't there and the killall has nothing to kill
to ruin the results). Second, it cleans up the chpl_pvm_fprintf call to send
the file descriptor (if one of the standard streams) to the launcher so the
parent node knows what to do with the buffer. If the stream isn't stdin,
stdout, or stderr, then just write to that file descriptor (likely a file).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15629 3a8e244f-b0f2-452b-bcba-4c88e055c3ca