Added -memleaks flag to parallel testing script.  To use this, specify a log file after the flag to store the final results.  Unlike with start_test, it is not appended but rather newly created.
Updated the testing scripts so that the execopts included with a test
via a .execopts or EXECOPTS file will be prepended to the execopts
passed to the start_test script.  This gives precedence to any
execopts flags passed to the testing scripts.

Updated the testing scripts to set an environment variable
CHPL_MEM_LEAK_TESTING to true when doing memory leak testing so that
we can use the skipif files to skip tests that are known to fail with
memory leak testing.  Added SKIPIF files for some test directories.
These were all added because of failures due to either the help
message changing (the configuration variable memLeaksLog is set) or
(2) a configuration variable error message changing (the error
messages change from not seeing an expected next argument to seeing
the wrong type of expected next argument).

Added a perl script util/analyzeMemLeaksLog.pl that takes a memory
leaks log as created with the memLeaksLog flags or -memleaks test
script flags and aggregates the results.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15633 3a8e244f-b0f2-452b-bcba-4c88e055c3ca