Interesting future: cloned forall body
This is an interesting future capturing a somewhat mysterious (to me
anyway) issue that Srinivas found -- for fairly simple forall loops
over distributed arrays, we seem to create two identical copies of the
loop body functions.  This isn't necessarily an error but it seems
like a potential cause of compilation inefficiency that is worth a
look.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15635 3a8e244f-b0f2-452b-bcba-4c88e055c3ca