This fix generates a better tag for PVM messages.
This fix generates a more random tag for PVM messages. In running some
multithreaded functions, given that the hash was a function of the localeID
and the thread, sometimes the same tag was created for PVM communications.
When these collided, often programs would exit with PVM error -5, PvmNoData.
However, by creating a more random number (seeded with the time and
including the hash and a lock_num incremented every time a PVM call is
invoked by taking out a lock), the likelihood of a collision is very minimal.
This fix also adds a little more information if there's a PVM error.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15640 3a8e244f-b0f2-452b-bcba-4c88e055c3ca