Forwarding vfprintf() calls to parent in PVM comm layer
In looking at PVM executions with Lee, we noticed that the --verbose
flag wasn't generating the normal message as each locale checks in.
Turns out that the chpl_msg() call uses vfprintf() which needs to
be forwarded to the parent process much like printf/fprintf do.
This check-in intercepts those calls in the same way and forwards
them using the same code Lee developed for fprintf().

This checkin also fixes a minor problem that I was having in the
launcher with respect to our re-definition of printf.  In retrospect,
the launcher shouldn't redirect any printfs, so I restructured
PVM's chplio_md.h so that it doesn't do anything deep for the launcher --
just #defines PRINTF_DEF as in the default chplio_md.h.  This cleaned
up the build for me.

I also patched up a couple of memory issues pointed out by valgrind.
Others main remain.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15642 3a8e244f-b0f2-452b-bcba-4c88e055c3ca