Added code to free the compiler-inserted _EndCount classes, which are used to synchronize cobegin, coforall, and sync statements, as well as the main program.
Fixed a few bugs that surfaced due to this new code.  The issue is
that after the waitEndCount call, we try to free the EndCount class,
but there may be more code to execute in the downEndCount call.  I
added a wait on e.i at the end of waitEndCount (which also waits for
e.b) because we have to update it after we signal e.b in downEndCount.

Another bug that surfaced is that even after waiting on e.i, we still
run into a write after free error due to the write to e.i in
downEndCount.  This is implemented via an assignment function that
signals a condition variable.  The result is then written back to e.i
(after the signal) so we can get into a situation where the other
thread frees e.i before we write the sync var reference back.
Fortunately, we don't have to write the sync var reference back
because it doesn't change in the assignment function.  To fix this, I
removed the return from the overloaded assignment function on sync and
single variables, and updated the compiler to remove the move back of
this reference.

The final write after free error showed up in the pthreads runtime
code.  It seems that we have to unlock the mutex after we signal the
condition variable and not before.

This commit ensures that we free _EndCount classes on all tests
(decreasing the aggregated leaked memory of running all tests by
between 5 and 10 MB).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15644 3a8e244f-b0f2-452b-bcba-4c88e055c3ca