These changes fix the random tag in PVM and launching on nodes with a daemon.
This fix contains two changes. First, the random tag returned by makeTag
was still generating the same tag in some circumstances, so the localeID is
incorporated into the mix. Second, the launcher was having trouble with
nodes if the previous run left the pvmd running. This change tries to run
the daemon, and if there was an error, it kills anything running on that node.
If it still fails, then the launch fails.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15654 3a8e244f-b0f2-452b-bcba-4c88e055c3ca