Fix to last night's no-local, gasnet, baseline regressions
This fix is similar to Steve's r15627 change: it explicitly removes
an on clause from a coforall loop if it's local in order to avoid
creating an additional thread.  In his case it was the coforall loop
over the Locales array; in mine it's the coforall over the Realms
array (requiring the loop body to be replicated four times now).
I'm going to do some quick checking to see whether we can optimize
away this problem automatically rather than manually to avoid these
code clones.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15665 3a8e244f-b0f2-452b-bcba-4c88e055c3ca