Fixed bug in my multi-realm checkin
As mentioned in my multi-realm checkin from yesterday, I didn't quite
have things working so that when using multiple nodes, the nodes were
mapped to the locales correctly.  This checkin fixes that bug by
correcting a primitive on clause in ChapelLocale.chpl -- it was using
the locale's ID relative to its realm rather than its global ID.  With
this change, running with CHPL_COMM = gasnet on a number of nodes
generates the correct output for a multi-realm execution of
hello-multirealm.chpl.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15667 3a8e244f-b0f2-452b-bcba-4c88e055c3ca