Fixed support for thisRealm
In the same way that we have a built-in identifier "here" to refer
to the current locale, the realms technical paper proposed a keyword
"thisRealm" to do the same for the current realm.  I had added support
for this in my previous checkin, but hadn't tested it and it turned
out not to work because I was only setting it up on locale 0 of each
realm.

This checkin fixes that and rewrites the hello-multirealm test to
use the here and thisRealm variables rather than the loop counters
(which is arguably a more robust and interesting test).

We also ended up storing a locale's realm as part of the locale's data
structure, so another way to get at this information would be through
a method, such as:

        here.realm

I think I like this slightly better than a new identifier "thisRealm",
both to reduce the lack of concepts and because a realm spans multiple
locales, so seems less ideal to privatize.  That said, I think it
makes sense to stick to our technical report as closely as possible
for this first year's work, so I wasn't ready to take thisRealm out
yet.  Good topic for future discussion, though.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15669 3a8e244f-b0f2-452b-bcba-4c88e055c3ca