Added support for specifying 0 locales within a realm
Both Lee and I had run into problems when specifying 0 locales
for a given realm -- something that seems reasonable to support.
This check-in permits 0 locales to be specified within a realm
so long as there is at least one locale in some realm.  Added
a test showcasing this capability -- zeroLocalesPerRealm.chpl

This check-in also has two other small changes:

- if 0 locales are specified in every realm, we used to generate
  our standard "use -nl or --numLocales to specify the number of
  locales" which isn't correct/appropriate for multi-realm
  executions.  Specialized this error message based on whether
  we're in a numRealms==1 or > 1 case, and factored it back into
  arg.c in a new routine called chpl_specify_locales_error() so 
  that all the communication layers could use this common logic
  and error message if desired.  Made the pvm, gasnet, mpi,
  and armci comm layers call this routine.

- changed a case that I missed earlier in changing a reference to
  'r' in hello-multirealm.chpl into 'thisRealm' because it seems
  more interesting/a better test.

------------------------------------------------------------------

Some narrative about this change follows, not required reading:

This change was interesting because the problem that we were seeing
was that for idioms like "on r" or "r.Locales(0)" where r is a realm
value generated by "coforall r in Realms", we were getting into
trouble if r had 0 locales per realm -- things would blow up in ugly
ways.  Digging into this, it wasn't too surprising that this was the
case -- we were trying to go to a locale that didn't exist so of
course something would go wrong.  But how to fix it?

I kept trying to find the place where we defined how to do an "on r"
before remembering that "on r" (or "on l" for a locale) is about as
elegant as can be -- we make sure to allocate the realm or locale
object on the locale in question, so it works using the identical
mechanism as anything else in the language.  So all that was necessary
was to not create the realm object for a realm that had 0 locales.
Easy enough.

But then the problem was that expressions like "on r" or "r.Locales(0)"
could result in nil pointer dereferences since now Realms(k) was never
initialized and so is a nil pointer.  This could be worked around using
idioms like:

       coforall r in Realms do
         if r then
           on r do

but this seemed really unfortunate -- forcing a module writer or user
to explicitly test to make sure their r object was valid before using
it.

The first fix that came to mind was to specify a standalone iterator
like "allRealms" that would wrap this logic into it:

     def allRealms {
       for r in Realms do
         if r then
           yield r;
     }

but this again felt a little lame -- it means that the user or module
author can't simply iterate over the Realms array, but has to
side-step it using this iterator.

So then I came up with the fix that I checked in: use a sparse domain
rather than a dense one to represent RealmSpace where a realm's index
would only be present if its number of locales was greater than zero.
This turned out to be an incredibly simple, single-module change that
fixed the entire problem (all the other modified files are due to the
two side issues mentioned above).  Thus the code changed from:

    const RealmSpace: domain(1) = [0..numRealms-1];

to:

    const RealmBox: domain(1) = [0..numRealms-1];
 
    const RealmSpace: sparse subdomain(RealmBox) = chpl_genNonEmptyRealms();

    def chpl_genNonEmptyRealms() {
      for i in RealmBox do
        if chpl_numLocales(i) > 0 then
          yield i;
    }

and suddenly 0 locale-per-realm runs work as expected with no other
changes to the modules, compiler, or runtime.  What a productive
language!

Potential downsides:

- iteration over a sparse domain/array can be more expensive than a
  dense one -- particularly given that in the common multirealm case,
  the sparse domain is likely to be fully populated.  I don't think
  this is a concern because it'll typically be a very small sparse
  array, and doing anything across multiple realms is going to be
  expensive enough that it should wash out any overhead of iterating
  over the sparse domain.  If this is an issue we can create a sparse
  domain data structure that uses a dense representation with a
  boolean "present/absent" bit to get us back to the dense iteration +
  conditional that would've been required otherwise.

- any compilation of a Chapel code will result in sparse data
  structures getting generated for the Realm module.  I don't see this
  as that big of a deal -- we shouldn't fear our own language features
  in writing standard modules -- but it will increase compile time so
  next I'm planning on working on a specialization of the realms array
  for single-realm executions to declare it to be a single-index dense
  arithmetic domain.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15677 3a8e244f-b0f2-452b-bcba-4c88e055c3ca