Specialization of RealmSpace
This checkin specializes the declaration of RealmSpace so that
when we're doing a single-realm run, it's just the 1D domain
[0..0] whereas on a multi-realm run, it's the sparse 1D domain
described in my previous commit.  The advantage of this is that
we don't end up generating a bunch of unused sparse domain/array
code in the common case of doing single-realm compilations which
would increase compile time without much benefit since the single
index would always be present by definition.

Implemented this using a similar trick that I used for the initial
realms check-in: created two ChapelRealmSpace modules, one that's
used for single-realm compilations, a second used for multi-realm
compilations.  The vast majority of the code remains in
ChapelRealm.chpl which is written in terms of ChapelRealm without
worrying about its definition.

Changed ChapelStandard.chpl to use this new module and filesToAST.cpp
to parse it.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15678 3a8e244f-b0f2-452b-bcba-4c88e055c3ca