Changes to make chpl_globals_registry more consistent across locales
In addition to the main change described below, this check-in also
changes the type of chpl_globals_registry from an array of char*
(char**) to an array of void* (void**).  I see this as a clarification
because there's nothing about the global registry that has to do with
characters and we don't do any byte-oriented pointer math on the
globals registry that would be simplified by treating it as a char**.
This change has the benefit of removing some casts to void* that
existed below, and should make the array easier to deal with within
gdb.

Between r13701 and this one, the variable chpl_globals_registry[] has
had a different definition on locale 0 than it has on locales
1..numLocales-1.  While this has worked, it has made the communication
code a bit more complicated than it needs to be and has led to a
reasonable amount of confusion when assuming that a variable with a
given name will play the same role on all locales.  This checkin makes
the variable play the same role on all locales.

To summarize how the world just was, given a module-scope Chapel
variable:

  var x: int;

the real 'x' would be stored on locale 0 and all locales (including
locale 0) would store a wide reference to x -- call it wx for the sake
of this discussion.  The chpl_globals_registry variable was (and still
is) used to initialize these wide references so that each locale's
wide reference address field (wx.addr) stores the correct address.

The way this was done is that locale 0 would store the address of x
(&x) in chpl_globals_registry[i] (where 'i' is an index that uniquely
refers to 'x') while locales 1..numLocales-1 would store &(wx.addr) in
the same slot of their copy of the array.  Then, each locale would
initialize its wx.addr by storing into *(chpl_globals_registry[i]) on
their locale the value stored at chpl_globals_registry[i] on locale 0.
This led to the asymmetry mentioned before -- get() calls would not
look symmetric in the source and destination addresses, pvm calls
would not pack/unpack in symmetric ways, the values you'd want to
print for chpl_globals_registry on locale 0 differed from those on
locales >1, etc.

The impetus for the asymmetry came about when we started supporting
segment fast for GASNet.  Segment fast can't access any remote address
-- only those on the shared heap -- so the &x values on locale 0 had
to be moved onto the shared heap (whereas previously,
chpl_globals_registry had been a statically declared array and
therefore not remotely accessible for segment fast).  This was done by
making chpl_globals_registry point to (alias) the start of the shared
heap.  However, it is possible to make this work without changing the
definition of chpl_globals_registry on each locale, so that's what
I've done in this change.

After this change, chpl_globals_registry[i] on each locale stores
&(wx.addr) such that the role of chpl_comm_broadcast_global_vars() can
be thought of as assigning *(chpl_globals_registry[i]) on all locales
the value of *(chpl_globals_registry[i]) on locale 0.  I dealt with
the segment fast issue above by making chpl_globals_registry no longer
alias the shared heap and simply point to the statically-allocated
array.  Then, rather than treating locale 0 differently than locale k
in CHPL_HEAP_REGISTER_GLOBAL_VAR(), we treat them the same and perform
a second copy of the pointer onto the start of the global heap on
locale 0 in a gasnet-specific (and more specifically, segment
fast-specific) way.  The hook for this is via the macro
CHPL_HEAP_REGISTER_GLOBAL_VAR_EXTRA() which can optionally be
specified by a comm layer in its chplcomm_md.h header file.  If
unspecified it is stubbed out by default.

(implementation note: while it was tempting to just put that logic
directly into this file or into the chplcomm_md.h header itself, we
don't currently #include "gasnet.h" in any of these files, so
referring to the seginfo_table variable was not possible without
dramatically changing the structure of the code, including the
different set of flags that we use to compile GASNet code).

I've updated the PVM and GASNet comm layers to reflect this change and
am checking in now in order to get the change to Lee sooner.  I have
not yet updated MPI because I don't have a way to test it (will work
with David on this), nor ARMCI because I'm not certain that it's
necessary (and because moving this patch between computers is a bit of
a pain).  ARMCI uses a different technique for initializing globals
which is why the chpl_globals_registry_static variable can't be used
in place of chpl_globals_registry as it used to be and now is in 4/5
of our comm layers.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15685 3a8e244f-b0f2-452b-bcba-4c88e055c3ca