Getting Sun build working again
The Sun is a potentially interesting platform for realms
work because (a) it has a different endianness than our
linux boxes and (b) it has a shared file system.  With
any luck, this checkin should re-enable the ability to
build on our Suns if CHPL_HOST_PLATFORM and CHPL_TARGET_PLATFORM
are set to sunos-old and if the non-module gcc is used.  I
chose the target name so that users with real Suns will not
stumble over our hacking around our ancient include files/libraries
if their Suns are up-to-date.  I suspect we will want to exclude
these directories and mention of this setting from the release.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15691 3a8e244f-b0f2-452b-bcba-4c88e055c3ca