Add a bool parameter "alias" to arithmetic domains and arrays. Any array declared over a domain with the alias parameter true also has the reindexed parameter true.
An array declared with an alias domain can alias another array's data,
reindexed to match the alias's domain. The shape of the alias domain must
match the shape of the array to alias. Use the '=>' operator to create an alias:
var D: domain(1, alias=true) = [1..5];
var A: [1..10] real, B: [D] real;
B => A[1..10 by 2];

Updated tests to expect the extra parameter in the domain types printed in
errors.

Updated BlockDist, MultiBlockDist and CMODist to include the alias parameter,
but not to use it in any significant way.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15692 3a8e244f-b0f2-452b-bcba-4c88e055c3ca