Added initial support for annotating types of communicated buffers
[One tangential change made by this commit is that I cleaned up some
of the code generation logic with respect to code vs. header files and
runtime code generation vs. user code generation.  The most important
change here was a small change to the open/closeCFile interface, but
the diff for codegen.cpp looks large because I renamed the FILE*s used
in fprintfs for clarity.]

This checkin results in two GASNet regressions that I haven't been
able to debug yet.  I decided to check in before fixing these in order
to get this code into Lee's hands sooner rather than later:

[Error matching program output for multilocale/deitz/linked/test_linked3c]
[Error matching program output for multilocale/diten/oneOrMoreLocales/assocArrayOfRemoteArithArray]

To date, all communications through chpl_comm_* routines have used a
void* pointer and number of bytes to specify data that needs to be
transferred between locales.  While this is sufficient for homogenous
platforms, it isn't for heterogeneous platforms in which endianness
conversions may be required.  This change lays the groundwork for
being able to support such typing of communicated buffers.

In particular, when a global compiler variable,
genCommunicatedStructures, is set to true, the compiler will create a
new output file in the generated code named _type_structure.c.
Currently, genCommunicatedStructures is only turned on if CHPL_COMM is
set to pvm (and even then it may not be used -- see below).  This file
contains an array named chpl_strucType[] that defines, for each
datatype communicated, a list of field types and offsets of those
fields such that a communication layer can send the data in a portable
manner such as PVM's pk*/upk* routines or XDR.  For each field, the
type of that field is encoded (using an enumeration defined in
chpltypes.h) as well as the offset of the field within the void*
buffer (currently incorrect for nested records).  Since the number of
fields per structure varies, a sentinel field with the value
{CHPL_TYPE_DONE, -1} is used to indicate the last field. The file also
contains an enumeration chpl_rt_types that enumerates all the
communicated types.  These enumerations are created by prepending
chpl_rt_type_id to the C name of the type in the generated code.

To communicate this structural information to a call like
chpl_comm_fork or chpl_comm_put, a chpl_rt_types enumeration is sent
into the routine in place of the normal size argument that's used for
the homogenous case.  This permits us to support both homogenous and
heterogeneous communication with a single interface.  Currently, the
decision as to which interpretation is used is based on a cpp macro
CHPL_COMM_HETEROGENEOUS which is off by default, for performance
reasons.  Thus, a to support both homogenous and heterogeneous
communications, a chpl_comm_layer will need to have two
implementations of the interface, one for when CHPL_COMM_HETEROGENEOUS
is set, the second for when it isn't.  We could support both
communication styles within a single compile by negating the
enumeration values to distinguish them from a size value, but I did
not take this approach in this checkin, figuring that we wouldn't want
to pay that runtime overhead.

Much of the challenge of making this change was modifying the
CHPL_COMM macros in chplcomm.h to accept and pass along the enumerated
values.  In particular, while some of the CHPL_COMM macros took a type
field that could be used to textually create the chpl_rt_types
enumeration, others did not, using expressions like sizeof(var) in
order to determine the size of the buffer to chpl_comm*() routines.
In these cases, I added additional fields to the macros to take a type
parameter and modified code generation to pass this type information
along.  To keep the macros simple and test that I was passing the
right types, I rewrote macros that used sizeof(var) to use
sizeof(type) instead.  This also permitted me to make the only change
required in chplcomm.h to support heterogeneous vs. homogenous
compilation to be the definition of a macro that either evaluates to a
chpl_rt_type_id* enumeration or a sizeof() expression:

#ifdef CHPL_COMM_HETEROGENEOUS
#define SPECIFY_SIZE(type) chpl_rt_type_id_##type
#else
#define SPECIFY_SIZE(type) sizeof(type)
#endif

Implemented the ability for a type to print out its structure through
new methods on the Type AST: codegenStructure() (defined on most
types) and codegenFieldStructure() (defined on ClassType only).

We determine types that will need to be communicated either during
parallel.cpp (for fork and fork_nb calls) or at codegen time (for
CHPL_COMM* macros) using a call, registerTypeToStructurallyCodegen()
which adds the type to a Map (used as a set).  Then at the end of
codegen we iterate over the Map twice, once to print out the field
information and a second time to create the enumeration of types.


Limitations that need to be addressed in future commits:

- offset values in _type_structure.c need to be corrected for nested
  records

- fix GASNet regressions

- I have a feeling that strings/wide strings are going to lead to the
  standard set of problems.  In particular, I put some chpl_error()
  calls in when trying to communicate them heterogeneously.

- I also have a feeling that I'm generating more types into
  _type_structure.c than are necessary -- in particular, for a simple
  code it seems like a much larger set of types than necessary, and
  some of them seem to have no fields.  This doesn't represent an
  incorrectness, merely an inefficiency (and one that will only occur
  for PVM compilations currently).

- currently, remote value forwarding seems to break the creation of
  the _type_structure.c file, so we disable remote value forwarding
  if genCommunicatedStructures is off.

- for some reason I had to special-case the _ddata classes in the
  generation of chpl_comm_fork calls in expr.cpp.  I couldn't figure
  this out in time to deal with it effectively.

- I added the registerTypeToStructurallyCodegen() call to passes.h
  because codegen.cpp doesn't have its own header file and this is the
  only routine that it exports other than codegen() itself.  I'm not
  particularly happy with this.

- the Map typesToStructurallyCodegen should be changed to a Vec used
  as a set since we don't use the value field.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15698 3a8e244f-b0f2-452b-bcba-4c88e055c3ca