Reworked domain reference counting in the modules.
Renamed BaseDom reference count field from _count to _domCnt$ in order to
better distinguish it from a planned reference count field in BaseArr (which
will replace the reference count field in ddata).

Removed call of List.destroy on BaseDom._arrs since it is no longer necessary.
It was presumably used to delete the last ListNode which is now deleted when
the array is deleted.

Refactored destroyDom so that the call to delete the domain class is done at
the call site (now in the domain and array record destructors).  This
eliminates the need to write "delete this".  Unified the two versions of this
call, one of which took an array and was before defined in the
DefaultArithmeticArr class and called in its destructor.  This should make this
code apply to all types of arrays.

Still to clean, I left a number of increments of _domCnt$ in the
DefaultArithmeticArr class.  Since this is the only array class in the standard
modules that supports array aliases, this is working now, but it should be
possible to refactor this code so that these increments happen in the array
wrapper record.  Updated the CMO distributed array test to increment these
counts as well.

This commit improves the aggregate memory leak information slightly (from 27.4%
leaked to 26.7% leaked) and most of the gains seem to come from eliminating
associative arrays and better memory management in the CG benchmarks.  This is
not a completely consistent improvement (just mostly).  For example, we now
leak some memory in test/arrays/diten/arrayAliasing.chpl.  I'm going to look
into that briefly.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15699 3a8e244f-b0f2-452b-bcba-4c88e055c3ca