This change begins to incorporate heterogeneous chplType PVM comm send/recvs.
This change begins to incorporate into the PVM comm layer the changes
necessary to implement the chplType interface. Given that sz now has a
double meaning, if CHPL_COMM_HETEROGENEOUS is set, we can do different
things using sz.

In the homogeneous case, just do what has always been done (sz is the size
of the buffer).

In the heterogeneous case, the value now represents an index into the
chpl_structType generated table (in _type_structure.c). This represents a
chplType and an offset. For example, if we want to send something that is
generated as chpl___class_localson_fn, the code now sends a CHPL_TYPE_int32_t,
which is represented in PVM as an int. It then sends a
CHPL_TYPE_CLASS_REFERENCE, which is just a pointer. PVM translates this
based on the sizeof (void *) and sends it via the appropriate mechanism.

Some other clean up is done here, too.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15705 3a8e244f-b0f2-452b-bcba-4c88e055c3ca