Refactored where the increments of _domCnt$ occur.  This field of BaseDom is no longer accessed in DefaultArithmeticArr; it is incremented and accessed exclusively in the ChapelArray and ChapelDistribution.  This improves the reference counting approach in general since we don't have to deal with the reference counts internal to a specific distributed array.  Not only does this make it easier to write new distributed array classes, but the overall performance is better since we don't have to reference count temporary array classes as built up in the reallocate method.  Removed an odd increment of the domain reference count in the reallocate method (where it was incremented by 2).  This is no longer necessary, though it is unclear to me whether it was even before this commit.
This commit improves the aggregate memory leakage slightly (from
1.12GB memory leaked in all tests to 1.01GB).  Looking at parentheses
is misleading because the total allocated memory increased by 0.4GB
(8.6%), thus helping the percentages (from 25.9% leaked to 21.4%
leaked).  This increase is solely from an extra domain copy in the
slicing interface.  We'll want to optimize this general idiom in the
future.

This commit also improves the memory leakage of the test
arrays/diten/arrayAliasing.chpl that slipped in r15699 due to a more
uniform approach to reference counting.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15706 3a8e244f-b0f2-452b-bcba-4c88e055c3ca