Fixed a seemingly incorrect type argument to CHPL_COMM_WIDE_ARRAY_GET()
In debugging some output in _type_structure.c for hello-multirealm.chpl
when compiling with -snumRealms>1, we seemed to be trying to communicate
a reference to an int32 (_ref_int32_t) which broke my code in type.cpp
by trying to generate a reference to the _val type of a non-structure.
Steve thought it was puzzling that we would try to communicate this type,
so I looked into the code generation for the macro in question
(CHPL_COMM_WIDE_ARRAY_GET()), did some pattern matching against 
CHPL_COMM_WIDE_ARRAY_GET_VALUE() and came up with an implementation that
seemed more correct, and made the error go away.  I'll code review this
with Steve next week to make sure that it makes sense.

In addition to making the above case compile better, this seems to pass
all the comm=none test and all of the release/examples + multilocale
tests on GASNet, except that test_linked3.chpl is failing again.
This makes me wonder whether CHPL_COMM_WIDE_ARRAY_GET() will need to
take three type arguments like CHPL_COMM_WIDE_ARRAY_GET_VALUE() did,
since it was failing similarly this morning.  My next step will be to
look into that.





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15711 3a8e244f-b0f2-452b-bcba-4c88e055c3ca