Fixing last night's regressions
Apparently my testing based on examples/ and multilocale/ made me
overconfident last night -- my changing of the type argument was
definitely a step in the right direction and most of last night's
failures demonstrated that apparently sending _ref_int32_t to a
CHPL_COMM routine was a reasonable thing to do once I dug into them
in a bit more detail.  Weird.

So, this change backs out that change to the argument and adds a
case to my code to generate structural annotations to handle
references to primitive types like _ref_int32_t.

This change also changes the way the size field is generated for
chpl_comm_fork* calls so that it uses the SPECIFY_SIZE macro
like the CHPL_COMM* calls.  This unifies things a bit more and
simplifies code generation.  This was one place where I had to
add an extra underscore by hand in order to get things to work,
though I'm not entirely certain why.  Something for me to look
at with Steve next week.

Tested this on examples/ multilocale/ and the failing HPCC
directories from last night, and it seems solid.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15712 3a8e244f-b0f2-452b-bcba-4c88e055c3ca