This change updates the PVM launcher to allow for different file names.
This change updates the PVM launcher to allow for different file names. The
net effect of this change is little; if the user calls his app something like
./bin/linux64/chap_linux64, and a remote node is of type darwin, the spawn
will occur to ./bin/darwin/chap_darwin_real.

An easy future fix would include allowing the app to be run from the actual
directory where the app resides. In the above case, if the current working
directory were $CHPL_HOME/bin/linux64, and the user calls chap_linux64, the
launcher would fail (launching $CHPL_HOME/bin/linux64/chap_darwin_real).

Totally different directories fail, as well, and the only way to allow this
would be a config file or environment variable.

This commit was done to get something reasonable working for the deadline on
Tuesday, but could be easily expanded. Note that some debug is left in, though
commented out.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15713 3a8e244f-b0f2-452b-bcba-4c88e055c3ca