Refactored reference counting code related to arrays.  Removed sync int field from internal data vector class in default arithmetic array class.  Removed all reference counting code from the default arithmetic array.
Added code to BaseArr class and _array wrapper record to use the sync
int and array alias field in BaseArr.

Added a new method with an old name (changed the meaning of this
method) called destroyData.  This function is called before an array
class is destroyed if there are no more aliases to this array class.
A default version in BaseArr does nothing.

This commit adversely impacts the aggregate memory leakage slightly
(from 1.01GB memory leaked in all tests to 1.11GB) because some
aliases are apparently not freed.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15717 3a8e244f-b0f2-452b-bcba-4c88e055c3ca