Disabled error checking against returning locally scoped arrays in variable functions.  The compiler now allows locally scoped arrays to be returned in variable functions.  My reasoning is that we allow non-variable functions to return arrays (and we return them by reference even if they are local).  In addition, we've always allowed slices of local arrays to returned in both variable and non-variable functions.  This relaxation should also allow us to remove the "valid var" pragma, a compiler thorn.
Added a new function to the array wrapper record called newAlias that
is used whenever the => operator is used in a variable declaration.
This makes all cases that use => on non-arrays result in an error.

Moved most of the code dealing with destroying arrays from the array
wrapper record to the BaseArr class.  This allows for recursion and
makes the freeing of aliases more complete.

Added a function getBaseDom() that returns the dom field of an array
class to all array class implementations.  This may be something that
we really just want to have the compiler create eventually.  This
function is used to destroy domains of array aliases in the function
in BaseArr.  It requires dynamic dispatch (though we can avoid that if
we are willing to assume that the 'dom' field will always be first).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15724 3a8e244f-b0f2-452b-bcba-4c88e055c3ca